edit: analog pin source now outputs an int

+++ src/sources/arduino/analogPinSource.hpp
@@ -6,7 +6,7 @@
 
 namespace rheo {
 
-  source_fn<float> analogPinSource(int pin, uint8_t resolutionBits = 10) {
+  source_fn<int> analogPinSource(int pin, uint8_t resolutionBits = 10) {
     return [pin, resolutionBits](push_fn<float> push, end_fn _) {
       return [pin, resolutionBits, push]() {
         analogReadResolution(resolutionBits);
